refactor(kernel): centralize BoxFuture alias in nectar-tasks
Delete the duplicate BoxFuture definition in nectar-kernel/future.rs;
nectar-tasks is the single home for the alias. nectar-kernel now
re-exports nectar_tasks::BoxFuture, and its unsync feature forwards to
nectar-tasks/unsync so the Send/non-Send flip stays in lockstep
(multi_thread is keyed per-crate on CARGO_FEATURE_UNSYNC).

Consumer import sites keep resolving nectar_kernel::BoxFuture via the
re-export; they migrate to nectar_tasks directly as part of #610's
governor rewire, avoiding double churn.

Closes #611
